relocate lowerzkcnative package
