+++ pkg/zkc/vm/internal/memory/wom.go
@@ -34,3 +34,18 @@ func (p *WriteOnce[W]) Read(frame []W, address []register.Id, data []register.Id
 func (p *WriteOnce[W]) Contents() []W {
 	return p.data
 }
+
+// NewWriteOnce constructs an empty write-once memory.
+func NewWriteOnce[W util.Uinter64](name string, public bool, registers []register.Register) InputOutput[W] {
+	var kind Kind
+	//
+	if public {
+		kind = PUBLIC_WRITE_ONCE_MEMORY
+	} else {
+		kind = PRIVATE_WRITE_ONCE_MEMORY
+	}
+	//
+	return &WriteOnce[W]{
+		StaticArray: NewStaticArray[W](name, kind, registers),
+	}
+}
