+++ pkg/zkc/vm/word.go
@@ -56,8 +56,7 @@ type Uint = word.Uint
 // Uint64 initialises a given word with a 64bit value.  This will panic if the
 // given value exceeds the available bandwidth of the word in question.
 func Uint64[W Word[W]](val uint64) W {
-	var w W
-	return w.SetUint64(val)
+	return word.Uint64[W](val)
 }
 
 // ============================================================================
